Set alpha=Inf as the default for testEmptyDrops.
This is because a finite alpha is not universally safer; despite
representing an overdispersed multinomial, using it to compute p-values
for a multinomial-distributed vector will actually be anticonservative!

This means that an inaccurate alpha cannot be brushed under the carpet.
Previously, I was trusting that any finite alpha would be safer than
alpha=Inf, but it seems that this is not true. We can't just ignore poor
estimates of alpha that we get from the assumed-ambient counts.

So, we just keep it simple and default alpha=Inf, which is exactly
correct for the multinomial case (which should hopefully be most cases).
This also reduces runtime and aligns with the CellRanger defaults.