Set alpha=Inf as the default for testEmptyDrops. (#118)
This is because a finite alpha is not universally safer; despite
representing an overdispersed multinomial, using it to compute p-values
for a multinomial-distributed vector will actually be anticonservative!

This means that an inaccurate alpha cannot be brushed under the carpet.
Previously, I was trusting that any finite alpha would be safer than
alpha=Inf, but it seems that this is not true. We can't just ignore poor
estimates of alpha that we get from the assumed-ambient counts.

So, we keep it simple and set the default to alpha=Inf, which is exactly
correct for the multinomial case, i.e., independent sampling of molecules into
droplet (which should represent most scenarios, hopefully). This also reduces
runtime and aligns with the CellRanger defaults.