Improve the quality of the knee point detection, again. (#124)
This still uses the "maximum distance from the line" approach but with a
hard-coded gradient of -1. The idea is to avoid the line having a shallow
gradient that shifts the estimated knee left from its visual location.
The value of -1 is basically arbitrary but works well for most datasets.

In the unlikely case where the fixed gradient line fails to identify a knee
point (because the curve is so shallow that its strongest gradient is not more
extreme than -1), we fall back to the previous method of drawing the line
from the inflection point to the trimmed start of the curve.

We also refine the reported inflection point to be within 1 log10-unit away
from the knee point. This is mostly for aesthetic purposes and ensures that the
appropriate location is identified if multiple inflection points are present.

'fit.bounds' is now a deprecated parameter as there's no more spline fitting.